+++ src/tuples.c
@@ -15,9 +15,9 @@
 
 t_tuple	tuple(t_cordinate x, t_cordinate y, t_cordinate z, t_cordinate type)
 {
-	float	*tpe;
+	double	*tpe;
 
-	tpe = malloc(sizeof(float) * 4);
+	tpe = malloc(sizeof(double) * 4);
 	if (!tpe)
 		return (NULL);
 	tpe[0] = x;
@@ -44,7 +44,7 @@ t_vector	normalize(t_tuple t)
 	return (scalar_multiplication(t, 1 / magnitude(t)));
 }
 
-float	dot_product(t_vector v1, t_vector v2)
+double	dot_product(t_vector v1, t_vector v2)
 {
 	return (v1[0] * v2[0] + v1[1] * v2[1] + v1[2] * v2[2]);
 }
